+++ modules/preboot/src/browser/listen/listen_by_selectors.ts
@@ -1,6 +1,7 @@
-import {PrebootRef} from '../../interfaces/preboot_ref';
+import {AppState} from '../../interfaces/preboot_ref';
 import {ListenStrategy} from '../../interfaces/strategy';
 import {NodeEvent} from '../../interfaces/event';
+import * as app from '../app'
 
 /**
  * This listen strategy uses a list of selectors maped to events. For example:
@@ -9,7 +10,7 @@ import {NodeEvent} from '../../interfaces/event';
  *      'button': ['click']
  *    }
  */
-export function getNodeEvents(preboot: PrebootRef, strategy: ListenStrategy): NodeEvent[] {
+export function getNodeEvents(appstate:AppState, strategy: ListenStrategy): NodeEvent[] {
   let nodeEvents = [];
   let eventsBySelector = strategy.eventsBySelector || {};
   let selectors = Object.keys(eventsBySelector);
@@ -17,7 +18,7 @@ export function getNodeEvents(preboot: PrebootRef, strategy: ListenStrategy): No
   // loop through selectors
   for (let selector of selectors) {
     let events = eventsBySelector[selector];
-    let elems = preboot.dom.getAllAppNodes(selector);
+    let elems = app.getAllAppNodes(appstate, selector);
 
     // if no elems, go to next iteration in for loop
     if (!elems) { continue; }
